empirical.controls {sva}	R Documentation

A function for estimating the probability that each gene is an empirical control

Description

This function uses the iteratively reweighted surrogate variable analysis approach to estimate the probability that each gene is an empirical control.

Usage


empirical.controls(dat, mod, mod0 = NULL, n.sv, B = 5, type = c("norm",
  "counts"))

Arguments

dat
The transformed data matrix with the variables in rows and samples in columns

mod
The model matrix being used to fit the data

mod0
The null model being compared when fitting the data

n.sv
The number of surogate variables to estimate

B
The number of iterations of the irwsva algorithm to perform

type
If type is norm then standard irwsva is applied, if type is counts, then the moderated log transform is applied first

Value

pcontrol A vector of probabilites that each gene is a control.

Examples


library(bladderbatch)
data(bladderdata)
dat <- bladderEset[1:5000,]

pheno = pData(dat)
edata = exprs(dat)
mod = model.matrix(~as.factor(cancer), data=pheno)

n.sv = num.sv(edata,mod,method="leek")
pcontrol <- empirical.controls(edata,mod,mod0=NULL,n.sv=n.sv,type="norm")


[Package sva version 3.32.1 Index]