psva {sva}	R Documentation

A function for estimating surrogate variables with the two step approach of Leek and Storey 2007

Description

This function is the implementation of the two step approach for estimating surrogate variables proposed by Leek and Storey 2007 PLoS Genetics. This function is primarily included for backwards compatibility. Newer versions of the sva algorithm are available through sva, svaseq, with low level functionality available through irwsva.build and ssva.

Usage


psva(dat, batch, ...)

Arguments

dat
The transformed data matrix with the variables in rows and samples in columns

batch
A factor variable giving the known batch levels

...
Other arguments to the sva function.

Value

psva.D Data with batch effect removed but biological heterogeneity preserved

Author(s)

Elana J. Fertig

Examples





library(bladderbatch)
library(limma)
data(bladderdata)
dat <- bladderEset[1:50,]

pheno = pData(dat)
edata = exprs(dat)
batch = pheno$batch
batch.fac = as.factor(batch)

psva_data <- psva(edata,batch.fac)


[Package sva version 3.32.1 Index]