sva_network {sva}	R Documentation

A function to adjust gene expression data before network inference

Description

This function corrects a gene expression matrix prior to network inference by returning the residuals after regressing out the top principal components. The number of principal components to remove can be determined using a permutation-based approach using the "num.sv" function with method = "be"

Usage


sva_network(dat, n.pc)

Arguments

dat
The raw gene expression data matrix (with variables in rows and samples in columns)

n.pc
The number of principal components to remove

Value

dat.adjusted Cleaned gene expression data matrix with the top prinicpal components removed

Examples


library(bladderbatch)
data(bladderdata)
dat <- bladderEset[1:5000,]

pheno = pData(dat)
edata = exprs(dat)
mod = model.matrix(~as.factor(cancer), data=pheno)

n.pc = num.sv(edata, mod, method="be")
dat.adjusted = sva_network(edata, n.pc)


[Package sva version 3.32.1 Index]